broken extender property on embedded services (#583)
* fix extender test setting wrong type,
and add new extender test that does not use 'as' keyword in embed.

* added test for issue 553

* minor fix to extender.ol test

* added that service embeded using 'embed' now loads interface extenders by adding them to the program in service loader.

* extender?.ol tests - no protocol for local communication required

* extender.ol test - fix the caller's interface

In this case "embed EmbedderService as EmbedderService" will never work since
the original interface (NothingServiceInterface) does not contain the token.
This causes Jolie's type checker to fail.

That said we are required to define the output port manually with an extended
interface (NothingServiceInterfaceWithToken) as shown in extender2.ol.

* wip interfaces extender

* extender.ol test - courier order + no protocol for local communication

* extender.ol test - fix the caller's interface

In this case "embed EmbedderService as EmbedderService" will never work since
the original interface (NothingServiceInterface) does not contain the token.
This causes Jolie's type checker to fail.

* OutputPortInfo carries the InterfaceExtender

and it is evaluated in OOIT, that is, applying the extender to the request/response types.

* fix pmd

* update test

* tidying code

* Remove duplicate Interface check

---------

Co-authored-by: Niels Erik Jepsen <nijep22@student.sdu.dk>
Co-authored-by: Claudio Guidi <cguidi@imolainformatica.it>
Co-authored-by: Narongrit Unwerawattana <narongrit.kie@gmail.com>